requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.1
google-cloud-texttospeech==2.16.3
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1